Moved social media analytics to dashboard
