+++ ai-brand-automator-frontend/src/app/dashboard/page.tsx
@@ -3,6 +3,7 @@
 import { OverviewCards } from '@/components/dashboard/OverviewCards';
 import RecentActivity from '@/components/dashboard/RecentActivity';
 import { QuickActions } from '@/components/dashboard/QuickActions';
+import SocialAnalytics from '@/components/dashboard/SocialAnalytics';
 import { useAuth } from '@/hooks/useAuth';
 
 export default function DashboardPage() {
@@ -18,6 +19,13 @@ export default function DashboardPage() {
         <div className="grid grid-cols-1 lg:grid-cols-3 gap-6">
           <div className="lg:col-span-2">
             <OverviewCards />
+            
+            {/* Social Analytics Section */}
+            <div className="mt-6">
+              <h2 className="text-xl font-heading font-semibold text-white mb-4">Social Media Analytics</h2>
+              <SocialAnalytics />
+            </div>
+            
             <div className="mt-6">
               <RecentActivity />
             </div>
